% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_curve_search.R
\name{mk_curve}
\alias{mk_curve}
\title{FUnction to get the curve attributes.}
\usage{
mk_curve(query_list = NULL, method = "get", token = NULL)
}
\arguments{
\item{query_list}{Named list with search params. If the get method is
used the query_list must uniquely identify a curve.}

\item{method}{Method: get, search or access}

\item{token}{Token}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(mkonline2)

mk_curve(query_list = list(query = "de"), method = "search")
mk_curve(query_list = list(id = 2), method = "get")
mk_curve(query_list = list(id = 2), method = "access")

}
