% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_curve_search.R
\name{mk_curve_search}
\alias{mk_curve_search}
\title{FUnction to get the curve attributes.}
\usage{
mk_curve_search(token = NULL, query_list = NULL)
}
\arguments{
\item{token}{Token}

\item{query_list}{Named list with search params.}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(mkonline2)

mk_curve_search(query_list = list(query = "de"))

}
