% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_token.R
\name{mk_token}
\alias{mk_token}
\title{Function to get a valid token for the API.}
\usage{
mk_token(client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{A client id from MK Online.}

\item{client_secret}{A client secret from MK Online.}
}
\description{
The token is used in Authorization header in each request to
the API.
}
