% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_curve_attributes.R
\name{ws_curve_attributes}
\alias{ws_curve_attributes}
\title{FUnction to get the curve attributes.}
\usage{
ws_curve_attributes(curve_attr, token = NULL)
}
\arguments{
\item{curve_attr}{The name of the curve attribute you want. The
current possibilities are: areas, categories, commodities}
}
\description{
FUnction to get the curve attributes.
}
