% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_series.R
\name{ws_series}
\alias{ws_series}
\title{FUnction to get the curve attributes.}
\usage{
ws_series(id, from = NULL, to = NULL, time_zone = NULL,
  ws_filter = NULL, ws_function = NULL, frequency = NULL,
  output_time_zone = NULL, token = NULL)
}
\arguments{
\item{id}{Curve id. Integer.}

\item{from}{character.}

\item{to}{Character.}

\item{time_zone}{Character.}

\item{ws_filter}{Character.}

\item{ws_function}{Character.}

\item{frequency}{Character.}

\item{output_time_zone}{Character.}

\item{token}{Token}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(wattsight2)

ws_series(curve_id = 2)

}
