% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_series_tagged.R
\name{ws_tagged}
\alias{ws_tagged}
\title{FUnction to get the curve attributes.}
\usage{
ws_tagged(id, tag = NULL, from = NULL, to = NULL, time_zone = NULL,
  ws_filter = NULL, ws_function = NULL, frequency = NULL,
  output_time_zone = NULL, token = NULL)
}
\arguments{
\item{id}{Curve id. Integer.}

\item{tag}{Time series tag. Array [string].}

\item{from}{Values from date (inclusive). Character.}

\item{to}{Values to date (exclusive). Character.}

\item{time_zone}{Change curve time_zone before filtering and frequency change. Character.}

\item{ws_filter}{Filter out parts of the time series. Array[string].}

\item{ws_function}{The aggregate/split function to use when changing frequency. Character.}

\item{frequency}{The required frequency of the output. Character.}

\item{output_time_zone}{Change curve time_zone after filtering and frequency change.}

\item{token}{Token}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(mkonline2)

}
