% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_series_tagged.R
\name{ws_tagged_tags}
\alias{ws_tagged_tags}
\title{FUnction to get the curve attributes.}
\usage{
ws_tagged_tags(id, token = NULL)
}
\arguments{
\item{id}{Curve id. Integer.}

\item{token}{Token}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(mkonline2)

}
