% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_series.R
\name{mk_series}
\alias{mk_series}
\title{FUnction to get the curve attributes.}
\usage{
mk_series(curve_id, token = NULL)
}
\arguments{
\item{curve_id}{Curve id. Integer.}

\item{token}{Token}
}
\description{
FUnction to get the curve attributes.
}
\examples{
library(mkonline2)

mk_series(curve_id = 2)

}
