% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_curve_attributes.R
\name{ws_curve_attributes}
\alias{ws_curve_attributes}
\title{FUnction to get the curve attributes.}
\usage{
ws_curve_attributes(curve_attr, token = NULL)
}
\arguments{
\item{curve_attr}{The name of the curve attribute you want. The
current possibilities are: areas, categories, commodities}
}
\description{
FUnction to get the curve attributes.
}
\examples{

ws_curve_attributes("areas")
ws_curve_attributes("categories")
ws_curve_attributes("commodities")
ws_curve_attributes("curve_types")
ws_curve_attributes("data_types")
ws_curve_attributes("frequencies")
ws_curve_attributes("functions")
ws_curve_attributes("sources")
ws_curve_attributes("stations")
# ws_curve_attributes("storage_types")
ws_curve_attributes("time_zones")
ws_curve_attributes("units")

}
